backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.2.3",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.4",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
lint = [
    "ruff>=0.8.4",
]
